fix(compaction): bind artifacts to durable source context
