+++ internal/conversation/flow/resolver_history.go
@@ -4,22 +4,14 @@ import (
 	"context"
 	"encoding/json"
 	"errors"
-	"fmt"
 	"log/slog"
 	"strings"
 	"time"
 
-	"github.com/google/uuid"
-	"github.com/jackc/pgx/v5/pgtype"
-	sdk "github.com/memohai/twilight-ai/sdk"
-
-	"github.com/memohai/memoh/internal/compaction"
 	"github.com/memohai/memoh/internal/contextfrag"
 	"github.com/memohai/memoh/internal/conversation"
-	"github.com/memohai/memoh/internal/db"
 	"github.com/memohai/memoh/internal/historyfrag"
 	messagepkg "github.com/memohai/memoh/internal/message"
-	pipelinepkg "github.com/memohai/memoh/internal/pipeline"
 	"github.com/memohai/memoh/internal/toolapproval"
 	"github.com/memohai/memoh/internal/userinput"
 )
@@ -342,647 +334,4 @@ func requiredMessagesBeforeCutoff(messages []historyfrag.HistoryRecord, cutoff i
 		}
 	}
 	return required
-}
-
-func historyContextFragsForMessages(messages []conversation.ModelMessage, records []historyfrag.HistoryRecord) []contextfrag.ContextFrag {
-	if len(messages) == 0 || len(records) == 0 {
-		return nil
-	}
-	frags := make([]contextfrag.ContextFrag, 0)
-	recordStart := 0
-	for i, msg := range messages {
-		if !looksLikeSummaryMessage(msg) {
-			continue
-		}
-		for j := recordStart; j < len(records); j++ {
-			record := records[j]
-			if record.Kind != contextfrag.KindConversationSummary || record.Coverage == nil {
-				continue
-			}
-			if !sameModelMessage(record.ModelMessage, msg) {
-				continue
-			}
-			frag := historyfrag.ToFrag(record)
-			frag.ID = fmt.Sprintf("message.%03d", i)
-			frag.Provenance.Index = i
-			frags = append(frags, frag)
-			recordStart = j + 1
-			break
-		}
-	}
-	return frags
-}
-
-func looksLikeSummaryMessage(msg conversation.ModelMessage) bool {
-	return strings.EqualFold(strings.TrimSpace(msg.Role), "user") &&
-		strings.HasPrefix(strings.TrimSpace(msg.TextContent()), "<summary>")
-}
-
-func sameModelMessage(a conversation.ModelMessage, b conversation.ModelMessage) bool {
-	return strings.EqualFold(strings.TrimSpace(a.Role), strings.TrimSpace(b.Role)) &&
-		string(a.Content) == string(b.Content)
-}
-
-func (r *Resolver) replaceCompactedMessages(ctx context.Context, sessionID string, scope contextfrag.Scope, messages []historyfrag.HistoryRecord) ([]historyfrag.HistoryRecord, error) {
-	if r.queries == nil {
-		return messages, nil
-	}
-	if strings.TrimSpace(sessionID) == "" {
-		// Sessionless (chat-scoped) loads have no session log list to draw from;
-		// resolve each in-window compact group individually.
-		return r.replaceRecentCompactedMessages(ctx, scope, messages)
-	}
-	frontier, err := r.loadActiveCompactionFrontier(ctx, scope.BotID, sessionID)
-	if err != nil {
-		return nil, err
-	}
-	if len(frontier.Artifacts) == 0 {
-		return messages, nil
-	}
-	messages = replaceCompactedHistoryRecordsWithResolver(messages, scope, frontier.Resolve, frontier.ResolveCoveredRef)
-	sessionSummaries := summaryRecordsFromArtifacts(missingCompactionArtifacts(messages, frontier.Artifacts), scope)
-	if len(sessionSummaries) > 0 {
-		messages = prependMissingCompactionSummaries(messages, sessionSummaries)
-	}
-	return r.refreshCompactedSummaryCoverage(ctx, messages, frontier.Resolve), nil
-}
-
-// missingCompactionArtifacts filters the active frontier down to summaries not
-// already represented in the loaded records.
-func missingCompactionArtifacts(messages []historyfrag.HistoryRecord, artifacts []compaction.Artifact) []compaction.Artifact {
-	seen := make(map[string]struct{}, len(messages))
-	for _, record := range messages {
-		if record.SourceKind == historyfrag.SourceCompactionLog {
-			if id := strings.TrimSpace(record.Ref.ID); id != "" {
-				seen[id] = struct{}{}
-			}
-		}
-	}
-	missing := make([]compaction.Artifact, 0, len(artifacts))
-	for _, artifact := range artifacts {
-		if _, ok := seen[artifact.ID]; ok {
-			continue
-		}
-		missing = append(missing, artifact)
-	}
-	return missing
-}
-
-func (r *Resolver) replaceRecentCompactedMessages(ctx context.Context, scope contextfrag.Scope, messages []historyfrag.HistoryRecord) ([]historyfrag.HistoryRecord, error) {
-	if r.queries == nil {
-		return messages, nil
-	}
-
-	compactGroups := make(map[string][]int) // compact_id -> indices
-	for i, m := range messages {
-		if m.CompactID != "" {
-			compactGroups[m.CompactID] = append(compactGroups[m.CompactID], i)
-		}
-	}
-	artifacts := make(map[string]compaction.Artifact)
-	frontiers := make(map[compaction.ArtifactOwner]compaction.ArtifactFrontier)
-	projection := compaction.NewArtifactProjection(r.queries)
-	for _, record := range messages {
-		owner := recordArtifactOwner(record, scope)
-		if owner.SessionID == "" {
-			continue
-		}
-		if _, loaded := frontiers[owner]; loaded {
-			continue
-		}
-		frontier, err := r.loadActiveCompactionFrontier(ctx, owner.BotID, owner.SessionID)
-		if err != nil {
-			return nil, err
-		}
-		frontiers[owner] = frontier
-		for _, artifact := range frontier.Artifacts {
-			artifacts[artifact.ID] = artifact
-		}
-	}
-	for compactID := range compactGroups {
-		owner, consistent := compactGroupOwner(messages, compactGroups[compactID], scope)
-		if !consistent {
-			if r.logger != nil {
-				r.logger.Warn("replaceCompactedMessages: compact group spans owners", slog.String("compact_id", compactID))
-			}
-			continue
-		}
-		if frontier, ok := frontiers[owner]; ok {
-			if artifact, resolved := frontier.Resolve(compactID); resolved {
-				artifacts[compactID] = artifact
-				artifacts[artifact.ID] = artifact
-			}
-			continue
-		}
-		artifact, err := projection.LoadActiveByID(ctx, compactID, owner)
-		if err != nil {
-			var lineageErr *compaction.LineageError
-			if !errors.As(err, &lineageErr) {
-				return nil, err
-			}
-			if r.logger != nil {
-				r.logger.Warn("replaceCompactedMessages: failed to load compaction artifact", slog.String("compact_id", compactID), slog.Any("error", err))
-			}
-			continue
-		}
-		artifacts[compactID] = artifact
-		artifacts[artifact.ID] = artifact
-	}
-	resolve := func(id string) (compaction.Artifact, bool) {
-		artifact, ok := artifacts[id]
-		return artifact, ok
-	}
-	resolveCovered := func(ref contextfrag.ContextRef) (compaction.Artifact, bool) {
-		var matched compaction.Artifact
-		for _, frontier := range frontiers {
-			artifact, ok := frontier.ResolveCoveredRef(ref)
-			if !ok {
-				continue
-			}
-			if matched.ID != "" && matched.ID != artifact.ID {
-				return compaction.Artifact{}, false
-			}
-			matched = artifact
-		}
-		seen := make(map[string]struct{})
-		for _, artifact := range artifacts {
-			if _, duplicate := seen[artifact.ID]; duplicate || !artifact.Covers(ref) {
-				continue
-			}
-			seen[artifact.ID] = struct{}{}
-			if matched.ID != "" && matched.ID != artifact.ID {
-				return compaction.Artifact{}, false
-			}
-			matched = artifact
-		}
-		return matched, matched.ID != ""
-	}
-
-	replaced := replaceCompactedHistoryRecordsWithResolver(messages, scope, resolve, resolveCovered)
-	return r.refreshCompactedSummaryCoverage(ctx, replaced, resolve), nil
-}
-
-func (r *Resolver) loadActiveCompactionFrontier(ctx context.Context, botID, sessionID string) (compaction.ArtifactFrontier, error) {
-	frontier, err := compaction.NewArtifactProjection(r.queries).LoadActiveSession(ctx, compaction.ArtifactOwner{BotID: botID, SessionID: sessionID})
-	if err != nil {
-		return compaction.ArtifactFrontier{}, err
-	}
-	for _, issue := range frontier.Issues {
-		if r.logger != nil {
-			r.logger.Warn("loadActiveCompactionFrontier: ignored invalid lineage", slog.String("issue", issue.Error()))
-		}
-	}
-	for _, artifact := range frontier.Artifacts {
-		if artifact.CoverageMalformed && r.logger != nil {
-			r.logger.Warn("loadActiveCompactionFrontier: malformed coverage requires legacy backfill", slog.String("compact_id", artifact.ID))
-		}
-	}
-	return frontier, nil
-}
-
-func recordArtifactOwner(record historyfrag.HistoryRecord, fallback contextfrag.Scope) compaction.ArtifactOwner {
-	return compaction.ArtifactOwner{
-		BotID:     firstNonEmpty(strings.TrimSpace(record.BotID), strings.TrimSpace(fallback.BotID)),
-		SessionID: firstNonEmpty(strings.TrimSpace(record.SessionID), strings.TrimSpace(fallback.SessionID)),
-	}
-}
-
-func compactGroupOwner(messages []historyfrag.HistoryRecord, indices []int, fallback contextfrag.Scope) (compaction.ArtifactOwner, bool) {
-	owner := compaction.ArtifactOwner{BotID: strings.TrimSpace(fallback.BotID), SessionID: strings.TrimSpace(fallback.SessionID)}
-	for _, index := range indices {
-		recordOwner := recordArtifactOwner(messages[index], contextfrag.Scope{})
-		if recordOwner.BotID != "" {
-			if owner.BotID != "" && owner.BotID != recordOwner.BotID {
-				return compaction.ArtifactOwner{}, false
-			}
-			owner.BotID = recordOwner.BotID
-		}
-		if recordOwner.SessionID != "" {
-			if owner.SessionID != "" && owner.SessionID != recordOwner.SessionID {
-				return compaction.ArtifactOwner{}, false
-			}
-			owner.SessionID = recordOwner.SessionID
-		}
-	}
-	return owner, true
-}
-
-func summaryRecordsFromArtifacts(artifacts []compaction.Artifact, scope contextfrag.Scope) []historyfrag.HistoryRecord {
-	if len(artifacts) == 0 {
-		return nil
-	}
-	records := make([]historyfrag.HistoryRecord, 0, len(artifacts))
-	for _, artifact := range artifacts {
-		records = append(records, artifact.HistoryRecord(scope))
-	}
-	return records
-}
-
-// refreshCompactedSummaryCoverage backfills legacy summaries that predate
-// persisted artifact coverage. New artifacts already carry immutable refs and
-// avoid this per-summary query.
-func (r *Resolver) refreshCompactedSummaryCoverage(ctx context.Context, messages []historyfrag.HistoryRecord, resolve func(string) (compaction.Artifact, bool)) []historyfrag.HistoryRecord {
-	if r.queries == nil {
-		return messages
-	}
-	for i := range messages {
-		record := &messages[i]
-		if record.SourceKind != historyfrag.SourceCompactionLog || record.Kind != contextfrag.KindConversationSummary {
-			continue
-		}
-		if artifact, ok := resolve(record.CompactID); ok && len(artifact.Coverage) > 0 {
-			continue
-		}
-		compactID, err := db.ParseUUID(record.CompactID)
-		if err != nil {
-			continue
-		}
-		coverage := contextfrag.NewSummaryCoverage(record.Ref, r.coveredRefsForCompact(ctx, compactID))
-		record.Coverage = &coverage
-	}
-	return messages
-}
-
-func (r *Resolver) coveredRefsForCompact(ctx context.Context, compactID pgtype.UUID) []contextfrag.ContextRef {
-	if r.queries == nil || !compactID.Valid {
-		return nil
-	}
-	rows, err := r.queries.ListMessageRefsByCompactID(ctx, compactID)
-	if err != nil {
-		if r.logger != nil {
-			r.logger.Warn("coveredRefsForCompact: failed to load compacted message refs", slog.String("compact_id", pgUUIDString(compactID)), slog.Any("error", err))
-		}
-		return nil
-	}
-	refs := make([]contextfrag.ContextRef, 0, len(rows))
-	for _, row := range rows {
-		ref, err := historyfrag.DBMessageIdentityRef(pgUUIDString(row.ID))
-		if err != nil {
-			if r.logger != nil {
-				r.logger.Warn("coveredRefsForCompact: skipped compacted message ref", slog.String("compact_id", pgUUIDString(compactID)), slog.Any("error", err))
-			}
-			continue
-		}
-		refs = append(refs, ref)
-	}
-	return refs
-}
-
-func prependMissingCompactionSummaries(messages []historyfrag.HistoryRecord, summaries []historyfrag.HistoryRecord) []historyfrag.HistoryRecord {
-	if len(summaries) == 0 {
-		return messages
-	}
-	seen := make(map[string]struct{}, len(messages))
-	for _, record := range messages {
-		if id := strings.TrimSpace(record.CompactID); id != "" {
-			seen[id] = struct{}{}
-		}
-		if record.SourceKind == historyfrag.SourceCompactionLog {
-			if id := strings.TrimSpace(record.Ref.ID); id != "" {
-				seen[id] = struct{}{}
-			}
-		}
-	}
-	missing := make([]historyfrag.HistoryRecord, 0, len(summaries))
-	for _, summary := range summaries {
-		id := strings.TrimSpace(summary.CompactID)
-		if id == "" {
-			id = strings.TrimSpace(summary.Ref.ID)
-		}
-		if id == "" {
-			continue
-		}
-		if _, ok := seen[id]; ok {
-			continue
-		}
-		seen[id] = struct{}{}
-		missing = append(missing, summary)
-	}
-	if len(missing) == 0 {
-		return messages
-	}
-	out := make([]historyfrag.HistoryRecord, 0, len(missing)+len(messages))
-	out = append(out, missing...)
-	out = append(out, messages...)
-	return out
-}
-
-func pgUUIDString(value pgtype.UUID) string {
-	if !value.Valid {
-		return ""
-	}
-	parsed, err := uuid.FromBytes(value.Bytes[:])
-	if err != nil {
-		return ""
-	}
-	return parsed.String()
-}
-
-func replaceCompactedHistoryRecords(messages []historyfrag.HistoryRecord, summaries map[string]string, scope contextfrag.Scope) []historyfrag.HistoryRecord {
-	artifacts := make(map[string]compaction.Artifact, len(summaries))
-	for compactID, summary := range summaries {
-		artifacts[compactID] = compaction.Artifact{ID: compactID, Summary: summary}
-	}
-	return replaceCompactedHistoryRecordsWithArtifacts(messages, artifacts, scope)
-}
-
-func replaceCompactedHistoryRecordsWithArtifacts(messages []historyfrag.HistoryRecord, artifacts map[string]compaction.Artifact, scope contextfrag.Scope) []historyfrag.HistoryRecord {
-	return replaceCompactedHistoryRecordsWithResolver(messages, scope, func(id string) (compaction.Artifact, bool) {
-		artifact, ok := artifacts[id]
-		return artifact, ok
-	}, nil)
-}
-
-func replaceCompactedHistoryRecordsWithResolver(
-	messages []historyfrag.HistoryRecord,
-	scope contextfrag.Scope,
-	resolveID func(string) (compaction.Artifact, bool),
-	resolveCovered func(contextfrag.ContextRef) (compaction.Artifact, bool),
-) []historyfrag.HistoryRecord {
-	type artifactGroup struct {
-		artifact compaction.Artifact
-		indices  []int
-		required bool
-	}
-	assignments := make([]string, len(messages))
-	groups := make(map[string]*artifactGroup)
-	for i, record := range messages {
-		artifact, ok := compaction.Artifact{}, false
-		if compactID := strings.TrimSpace(record.CompactID); compactID != "" {
-			artifact, ok = resolveID(compactID)
-		}
-		if !ok && resolveCovered != nil {
-			artifact, ok = resolveCovered(record.Ref)
-		}
-		if !ok || artifact.ID == "" || strings.TrimSpace(artifact.Summary) == "" {
-			continue
-		}
-		assignments[i] = artifact.ID
-		group := groups[artifact.ID]
-		if group == nil {
-			group = &artifactGroup{artifact: artifact}
-			groups[artifact.ID] = group
-		}
-		group.indices = append(group.indices, i)
-		group.required = group.required || record.Required
-	}
-	if len(groups) == 0 {
-		return messages
-	}
-
-	result := make([]historyfrag.HistoryRecord, 0, len(messages))
-	emitted := make(map[string]struct{}, len(groups))
-	for i, record := range messages {
-		artifactID := assignments[i]
-		if artifactID == "" {
-			result = append(result, record)
-			continue
-		}
-		group := groups[artifactID]
-		if _, seen := emitted[artifactID]; seen {
-			if group.required {
-				result = append(result, record)
-			}
-			continue
-		}
-		emitted[artifactID] = struct{}{}
-		artifact := group.artifact
-		if len(artifact.Coverage) == 0 {
-			artifact.Coverage = make([]compaction.CoveredSource, 0, len(group.indices))
-			for _, index := range group.indices {
-				artifact.Coverage = append(artifact.Coverage, compaction.CoveredSource{Ref: messages[index].Ref})
-			}
-		}
-		result = append(result, artifact.HistoryRecord(scope))
-		if group.required {
-			result = append(result, record)
-		}
-	}
-	return result
-}
-
-// buildMessagesFromPipeline assembles chat context from the DCP pipeline's
-// RenderedContext (RC) merged with assistant/tool turns (TR) from
-// bot_history_messages. This gives chat mode the same event-driven context
-// that discuss mode uses, replacing the legacy loadMessages path.
-func (r *Resolver) buildMessagesFromPipeline(ctx context.Context, req conversation.ChatRequest, contextTokenBudget int) []conversation.ModelMessage {
-	sessionID := strings.TrimSpace(req.SessionID)
-	if r.pipeline == nil || sessionID == "" {
-		return nil
-	}
-	rc := r.pipeline.GetRC(sessionID)
-	if len(rc) == 0 {
-		return nil
-	}
-
-	trs := r.loadTurnResponses(ctx, sessionID)
-
-	composed := pipelinepkg.ComposeContext(rc, trs, "")
-	if composed == nil {
-		return nil
-	}
-
-	messages := make([]conversation.ModelMessage, 0, len(composed.Messages))
-	for _, m := range composed.Messages {
-		contentJSON := m.RawContent
-		if len(contentJSON) == 0 {
-			var err error
-			contentJSON, err = json.Marshal(m.Content)
-			if err != nil {
-				continue
-			}
-		}
-		messages = append(messages, conversation.ModelMessage{
-			Role:    m.Role,
-			Content: contentJSON,
-		})
-	}
-
-	// Apply context token budget trimming to pipeline path as well.
-	if contextTokenBudget > 0 && len(messages) > 0 {
-		messages = trimPipelineMessagesByTokens(r.logger, messages, contextTokenBudget)
-	}
-
-	return messages
-}
-
-// trimPipelineMessagesByTokens trims pipeline-assembled messages to fit within
-// the context token budget using character-based estimation.
-func trimPipelineMessagesByTokens(log *slog.Logger, messages []conversation.ModelMessage, maxTokens int) []conversation.ModelMessage {
-	totalTokens := 0
-	cutoff := 0
-	for i := len(messages) - 1; i >= 0; i-- {
-		totalTokens += estimateMessageTokens(messages[i])
-		if totalTokens > maxTokens {
-			cutoff = i + 1
-			break
-		}
-	}
-
-	// Avoid orphaned tool messages at the cutoff boundary.
-	for cutoff < len(messages) && strings.EqualFold(strings.TrimSpace(messages[cutoff].Role), "tool") {
-		cutoff++
-	}
-
-	if cutoff > 0 && log != nil {
-		log.Info("trimPipelineMessagesByTokens: context trimmed",
-			slog.Int("total_messages", len(messages)),
-			slog.Int("estimated_tokens", totalTokens),
-			slog.Int("max_tokens", maxTokens),
-			slog.Int("kept_messages", len(messages)-cutoff),
-		)
-	}
-
-	return messages[cutoff:]
-}
-
-// loadTurnResponses loads recent assistant/tool messages from bot_history_messages
-// for use as the TR stream in pipeline-based context assembly.
-func (r *Resolver) loadTurnResponses(ctx context.Context, sessionID string) []pipelinepkg.TurnResponseEntry {
-	if r.messageService == nil {
-		return nil
-	}
-	since := time.Now().UTC().Add(-24 * time.Hour)
-	msgs, err := r.messageService.ListActiveSinceBySession(ctx, sessionID, since)
-	if err != nil {
-		r.logger.Warn("load TRs failed", slog.String("session_id", sessionID), slog.Any("error", err))
-		return nil
-	}
-	var trs []pipelinepkg.TurnResponseEntry
-	for _, m := range msgs {
-		entry, ok := pipelinepkg.DecodeTurnResponseEntry(m)
-		if !ok {
-			continue
-		}
-		trs = append(trs, entry)
-	}
-	return trs
-}
-
-// stripToolMessages removes bulky tool interactions from the context while
-// keeping ask_user calls and results. ask_user is conversation-visible: the
-// question and the user's answer are part of the chat semantics, not tool noise.
-func stripToolMessages(messages []conversation.ModelMessage) []conversation.ModelMessage {
-	filtered := make([]conversation.ModelMessage, 0, len(messages))
-	for _, m := range messages {
-		role := strings.TrimSpace(m.Role)
-		if strings.EqualFold(role, "tool") {
-			if kept := keepAskUserToolResultMessage(m); kept != nil {
-				filtered = append(filtered, *kept)
-			}
-			continue
-		}
-		// Remove assistant messages that only contain tool calls / reasoning with
-		// no visible text. Tool-call metadata may live either in ToolCalls or in
-		// structured content parts.
-		if strings.EqualFold(role, "assistant") && hasToolCallContent(m) {
-			stripped, ok := stripNonAskUserToolCalls(m)
-			if !ok {
-				continue
-			}
-			m = stripped
-		}
-		filtered = append(filtered, m)
-	}
-	return filtered
-}
-
-func stripNonAskUserToolCalls(message conversation.ModelMessage) (conversation.ModelMessage, bool) {
-	legacyToolCalls := keepAskUserLegacyToolCalls(message.ToolCalls)
-	text := strings.TrimSpace(message.TextContent())
-
-	keptParts := filterAssistantContextParts(modelMessageToSDKMessage(message).Content)
-	if len(keptParts) > 0 {
-		message = modelMessageFromSDKParts(sdk.MessageRoleAssistant, keptParts, message.Usage)
-		message.ToolCalls = legacyToolCalls
-		return message, true
-	}
-
-	message.ToolCalls = legacyToolCalls
-	if len(message.ToolCalls) > 0 {
-		if text != "" {
-			message.Content = conversation.NewTextContent(text)
-		}
-		return message, true
-	}
-	if text == "" {
-		return conversation.ModelMessage{}, false
-	}
-	message.Content = conversation.NewTextContent(text)
-	return message, true
-}
-
-func keepAskUserToolResultMessage(message conversation.ModelMessage) *conversation.ModelMessage {
-	if strings.EqualFold(strings.TrimSpace(message.Name), userinput.ToolNameAskUser) {
-		return &message
-	}
-	results := filterAskUserToolResults(modelMessageToSDKMessage(message).Content)
-	if len(results) == 0 {
-		return nil
-	}
-	message = modelMessageFromSDKParts(sdk.MessageRoleTool, results, message.Usage)
-	return &message
-}
-
-func keepAskUserLegacyToolCalls(calls []conversation.ToolCall) []conversation.ToolCall {
-	if len(calls) == 0 {
-		return nil
-	}
-	kept := make([]conversation.ToolCall, 0, len(calls))
-	for _, call := range calls {
-		if strings.EqualFold(strings.TrimSpace(call.Function.Name), userinput.ToolNameAskUser) {
-			kept = append(kept, call)
-		}
-	}
-	return kept
-}
-
-func filterAssistantContextParts(parts []sdk.MessagePart) []sdk.MessagePart {
-	if len(parts) == 0 {
-		return nil
-	}
-	kept := make([]sdk.MessagePart, 0, len(parts))
-	for _, part := range parts {
-		switch typed := part.(type) {
-		case sdk.ToolCallPart:
-			if strings.EqualFold(strings.TrimSpace(typed.ToolName), userinput.ToolNameAskUser) {
-				kept = append(kept, typed)
-			}
-		case sdk.ToolResultPart, sdk.ReasoningPart:
-			continue
-		case sdk.TextPart:
-			if strings.TrimSpace(typed.Text) != "" {
-				kept = append(kept, typed)
-			}
-		default:
-			kept = append(kept, part)
-		}
-	}
-	return kept
-}
-
-func filterAskUserToolResults(parts []sdk.MessagePart) []sdk.MessagePart {
-	if len(parts) == 0 {
-		return nil
-	}
-	kept := make([]sdk.MessagePart, 0, len(parts))
-	for _, part := range parts {
-		result, ok := part.(sdk.ToolResultPart)
-		if !ok {
-			continue
-		}
-		if strings.EqualFold(strings.TrimSpace(result.ToolName), userinput.ToolNameAskUser) {
-			kept = append(kept, result)
-		}
-	}
-	return kept
-}
-
-func modelMessageFromSDKParts(role sdk.MessageRole, parts []sdk.MessagePart, usage json.RawMessage) conversation.ModelMessage {
-	converted := sdkMessagesToModelMessages([]sdk.Message{{Role: role, Content: parts}})
-	if len(converted) == 0 {
-		return conversation.ModelMessage{Role: string(role)}
-	}
-	converted[0].Usage = usage
-	return converted[0]
 }
