+++ internal/conversation/flow/resolver_history_messages.go
@@ -1,0 +1,241 @@
+package flow
+
+import (
+	"context"
+	"encoding/json"
+	"log/slog"
+	"strings"
+	"time"
+
+	sdk "github.com/memohai/twilight-ai/sdk"
+
+	"github.com/memohai/memoh/internal/conversation"
+	pipelinepkg "github.com/memohai/memoh/internal/pipeline"
+	"github.com/memohai/memoh/internal/userinput"
+)
+
+// buildMessagesFromPipeline assembles chat context from the DCP pipeline's
+// RenderedContext (RC) merged with assistant/tool turns (TR) from
+// bot_history_messages. This gives chat mode the same event-driven context
+// that discuss mode uses, replacing the legacy loadMessages path.
+func (r *Resolver) buildMessagesFromPipeline(ctx context.Context, req conversation.ChatRequest, contextTokenBudget int) []conversation.ModelMessage {
+	sessionID := strings.TrimSpace(req.SessionID)
+	if r.pipeline == nil || sessionID == "" {
+		return nil
+	}
+	rc := r.pipeline.GetRC(sessionID)
+	if len(rc) == 0 {
+		return nil
+	}
+
+	trs := r.loadTurnResponses(ctx, sessionID)
+
+	composed := pipelinepkg.ComposeContext(rc, trs, "")
+	if composed == nil {
+		return nil
+	}
+
+	messages := make([]conversation.ModelMessage, 0, len(composed.Messages))
+	for _, m := range composed.Messages {
+		contentJSON := m.RawContent
+		if len(contentJSON) == 0 {
+			var err error
+			contentJSON, err = json.Marshal(m.Content)
+			if err != nil {
+				continue
+			}
+		}
+		messages = append(messages, conversation.ModelMessage{
+			Role:    m.Role,
+			Content: contentJSON,
+		})
+	}
+
+	// Apply context token budget trimming to pipeline path as well.
+	if contextTokenBudget > 0 && len(messages) > 0 {
+		messages = trimPipelineMessagesByTokens(r.logger, messages, contextTokenBudget)
+	}
+
+	return messages
+}
+
+// trimPipelineMessagesByTokens trims pipeline-assembled messages to fit within
+// the context token budget using character-based estimation.
+func trimPipelineMessagesByTokens(log *slog.Logger, messages []conversation.ModelMessage, maxTokens int) []conversation.ModelMessage {
+	totalTokens := 0
+	cutoff := 0
+	for i := len(messages) - 1; i >= 0; i-- {
+		totalTokens += estimateMessageTokens(messages[i])
+		if totalTokens > maxTokens {
+			cutoff = i + 1
+			break
+		}
+	}
+
+	// Avoid orphaned tool messages at the cutoff boundary.
+	for cutoff < len(messages) && strings.EqualFold(strings.TrimSpace(messages[cutoff].Role), "tool") {
+		cutoff++
+	}
+
+	if cutoff > 0 && log != nil {
+		log.Info("trimPipelineMessagesByTokens: context trimmed",
+			slog.Int("total_messages", len(messages)),
+			slog.Int("estimated_tokens", totalTokens),
+			slog.Int("max_tokens", maxTokens),
+			slog.Int("kept_messages", len(messages)-cutoff),
+		)
+	}
+
+	return messages[cutoff:]
+}
+
+// loadTurnResponses loads recent assistant/tool messages from bot_history_messages
+// for use as the TR stream in pipeline-based context assembly.
+func (r *Resolver) loadTurnResponses(ctx context.Context, sessionID string) []pipelinepkg.TurnResponseEntry {
+	if r.messageService == nil {
+		return nil
+	}
+	since := time.Now().UTC().Add(-24 * time.Hour)
+	msgs, err := r.messageService.ListActiveSinceBySession(ctx, sessionID, since)
+	if err != nil {
+		r.logger.Warn("load TRs failed", slog.String("session_id", sessionID), slog.Any("error", err))
+		return nil
+	}
+	var trs []pipelinepkg.TurnResponseEntry
+	for _, m := range msgs {
+		entry, ok := pipelinepkg.DecodeTurnResponseEntry(m)
+		if !ok {
+			continue
+		}
+		trs = append(trs, entry)
+	}
+	return trs
+}
+
+// stripToolMessages removes bulky tool interactions from the context while
+// keeping ask_user calls and results. ask_user is conversation-visible: the
+// question and the user's answer are part of the chat semantics, not tool noise.
+func stripToolMessages(messages []conversation.ModelMessage) []conversation.ModelMessage {
+	filtered := make([]conversation.ModelMessage, 0, len(messages))
+	for _, m := range messages {
+		role := strings.TrimSpace(m.Role)
+		if strings.EqualFold(role, "tool") {
+			if kept := keepAskUserToolResultMessage(m); kept != nil {
+				filtered = append(filtered, *kept)
+			}
+			continue
+		}
+		// Remove assistant messages that only contain tool calls / reasoning with
+		// no visible text. Tool-call metadata may live either in ToolCalls or in
+		// structured content parts.
+		if strings.EqualFold(role, "assistant") && hasToolCallContent(m) {
+			stripped, ok := stripNonAskUserToolCalls(m)
+			if !ok {
+				continue
+			}
+			m = stripped
+		}
+		filtered = append(filtered, m)
+	}
+	return filtered
+}
+
+func stripNonAskUserToolCalls(message conversation.ModelMessage) (conversation.ModelMessage, bool) {
+	legacyToolCalls := keepAskUserLegacyToolCalls(message.ToolCalls)
+	text := strings.TrimSpace(message.TextContent())
+
+	keptParts := filterAssistantContextParts(modelMessageToSDKMessage(message).Content)
+	if len(keptParts) > 0 {
+		message = modelMessageFromSDKParts(sdk.MessageRoleAssistant, keptParts, message.Usage)
+		message.ToolCalls = legacyToolCalls
+		return message, true
+	}
+
+	message.ToolCalls = legacyToolCalls
+	if len(message.ToolCalls) > 0 {
+		if text != "" {
+			message.Content = conversation.NewTextContent(text)
+		}
+		return message, true
+	}
+	if text == "" {
+		return conversation.ModelMessage{}, false
+	}
+	message.Content = conversation.NewTextContent(text)
+	return message, true
+}
+
+func keepAskUserToolResultMessage(message conversation.ModelMessage) *conversation.ModelMessage {
+	if strings.EqualFold(strings.TrimSpace(message.Name), userinput.ToolNameAskUser) {
+		return &message
+	}
+	results := filterAskUserToolResults(modelMessageToSDKMessage(message).Content)
+	if len(results) == 0 {
+		return nil
+	}
+	message = modelMessageFromSDKParts(sdk.MessageRoleTool, results, message.Usage)
+	return &message
+}
+
+func keepAskUserLegacyToolCalls(calls []conversation.ToolCall) []conversation.ToolCall {
+	if len(calls) == 0 {
+		return nil
+	}
+	kept := make([]conversation.ToolCall, 0, len(calls))
+	for _, call := range calls {
+		if strings.EqualFold(strings.TrimSpace(call.Function.Name), userinput.ToolNameAskUser) {
+			kept = append(kept, call)
+		}
+	}
+	return kept
+}
+
+func filterAssistantContextParts(parts []sdk.MessagePart) []sdk.MessagePart {
+	if len(parts) == 0 {
+		return nil
+	}
+	kept := make([]sdk.MessagePart, 0, len(parts))
+	for _, part := range parts {
+		switch typed := part.(type) {
+		case sdk.ToolCallPart:
+			if strings.EqualFold(strings.TrimSpace(typed.ToolName), userinput.ToolNameAskUser) {
+				kept = append(kept, typed)
+			}
+		case sdk.ToolResultPart, sdk.ReasoningPart:
+			continue
+		case sdk.TextPart:
+			if strings.TrimSpace(typed.Text) != "" {
+				kept = append(kept, typed)
+			}
+		default:
+			kept = append(kept, part)
+		}
+	}
+	return kept
+}
+
+func filterAskUserToolResults(parts []sdk.MessagePart) []sdk.MessagePart {
+	if len(parts) == 0 {
+		return nil
+	}
+	kept := make([]sdk.MessagePart, 0, len(parts))
+	for _, part := range parts {
+		result, ok := part.(sdk.ToolResultPart)
+		if !ok {
+			continue
+		}
+		if strings.EqualFold(strings.TrimSpace(result.ToolName), userinput.ToolNameAskUser) {
+			kept = append(kept, result)
+		}
+	}
+	return kept
+}
+
+func modelMessageFromSDKParts(role sdk.MessageRole, parts []sdk.MessagePart, usage json.RawMessage) conversation.ModelMessage {
+	converted := sdkMessagesToModelMessages([]sdk.Message{{Role: role, Content: parts}})
+	if len(converted) == 0 {
+		return conversation.ModelMessage{Role: string(role)}
+	}
+	converted[0].Usage = usage
+	return converted[0]
+}
