+++ internal/conversation/flow/resolver_history_pipeline_records_test.go
@@ -1,0 +1,101 @@
+package flow
+
+import (
+	"reflect"
+	"testing"
+
+	sdk "github.com/memohai/twilight-ai/sdk"
+
+	"github.com/memohai/memoh/internal/contextfrag"
+	"github.com/memohai/memoh/internal/conversation"
+	"github.com/memohai/memoh/internal/historyfrag"
+	messagepkg "github.com/memohai/memoh/internal/message"
+)
+
+func TestHistoryRecordPathPreservesLegacyResolverMessagePipeline(t *testing.T) {
+	t.Parallel()
+
+	assistantToolCallSDK := sdk.Message{
+		Role: sdk.MessageRoleAssistant,
+		Content: []sdk.MessagePart{
+			sdk.ToolCallPart{
+				ToolCallID: "call-1",
+				ToolName:   "lookup",
+				Input:      map[string]any{"q": "memoh"},
+			},
+		},
+	}
+	assistantToolCall := sdkMessagesToModelMessages([]sdk.Message{assistantToolCallSDK})[0]
+	toolResultSDK := sdk.ToolMessage(sdk.ToolResultPart{
+		ToolCallID: "call-1",
+		ToolName:   "lookup",
+		Result:     "tool result",
+	})
+	toolResult := sdkMessagesToModelMessages([]sdk.Message{toolResultSDK})[0]
+	rows := []messagepkg.Message{
+		dbHistoryRow(t, "row-compact-user", "user", conversation.NewTextContent("old compacted user"), func(msg *messagepkg.Message) {
+			msg.CompactID = "compact-ok"
+		}),
+		dbHistoryRow(t, "row-compact-assistant", "assistant", conversation.NewTextContent("old compacted assistant"), func(msg *messagepkg.Message) {
+			msg.CompactID = "compact-ok"
+		}),
+		dbHistoryRow(t, "row-missing-summary", "user", conversation.NewTextContent("missing summary body"), func(msg *messagepkg.Message) {
+			msg.CompactID = "compact-missing"
+		}),
+		dbHistoryRow(t, "row-current", "user", conversation.NewTextContent("already persisted current"), func(msg *messagepkg.Message) {
+			msg.SessionID = "sess-1"
+			msg.ExternalMessageID = "msg-current"
+			msg.Platform = "telegram"
+			msg.SenderChannelIdentityID = "sender-1"
+		}),
+		{
+			ID:      "row-plain",
+			BotID:   "bot-1",
+			Role:    "user",
+			Content: conversation.NewTextContent("plain string content"),
+		},
+		dbHistoryRow(t, "row-tool-call", "assistant", mustRawJSON(t, assistantToolCall), nil),
+		dbHistoryRow(t, "row-tool-result", "tool", mustRawJSON(t, toolResult), nil),
+	}
+
+	records := make([]historyfrag.HistoryRecord, 0, len(rows))
+	for _, row := range rows {
+		record, err := historyfrag.FromDBMessage(row, historyfrag.ScopeFallback{ChatID: "chat-1"})
+		if err != nil {
+			t.Fatalf("FromDBMessage(%s): %v", row.ID, err)
+		}
+		records = append(records, record)
+	}
+	records = dedupePersistedCurrentUserMessage(records, conversation.ChatRequest{
+		UserMessagePersisted:    true,
+		SessionID:               "sess-1",
+		ExternalMessageID:       "msg-current",
+		CurrentChannel:          "telegram",
+		SourceChannelIdentityID: "sender-1",
+	})
+	records = replaceCompactedHistoryRecords(records, map[string]string{"compact-ok": "condensed"}, contextfrag.Scope{})
+	got, tokens := trimMessagesByTokens(nil, records, 0)
+
+	want := []conversation.ModelMessage{
+		{Role: "user", Content: conversation.NewTextContent("<summary>\ncondensed\n</summary>")},
+		{Role: "user", Content: conversation.NewTextContent("missing summary body")},
+		{Role: "user", Content: conversation.NewTextContent("plain string content")},
+		assistantToolCall,
+		toolResult,
+	}
+	if !reflect.DeepEqual(got, want) {
+		t.Fatalf("history pipeline payload mismatch:\ngot  %#v\nwant %#v", got, want)
+	}
+	if tokens == 0 {
+		t.Fatal("history pipeline should report estimated tokens for retained records")
+	}
+
+	repaired := repairToolCallClosures(sanitizeMessages(got), syntheticToolClosureError)
+	assertSameJSON(t, modelMessagesToSDKMessages(nonNilModelMessages(repaired)), []sdk.Message{
+		sdk.UserMessage("<summary>\ncondensed\n</summary>"),
+		sdk.UserMessage("missing summary body"),
+		sdk.UserMessage("plain string content"),
+		assistantToolCallSDK,
+		toolResultSDK,
+	})
+}
